// src/app/modules/modulo-sim/models/tx-sim-digestivo.model.ts
export class TxSIMDigestivoModel {
    idSIMDigestivo: number;
    idSIM: number;
    ave: number;
    duademo: number;
    yeyuno: number;
    lleon: number;
    ciegos: number;
    tonsillas: number;
    higados: number;
    molleja: number;
    proventriculo: number;
    flgGradoLesion: boolean;

    constructor() {
        this.idSIMDigestivo = 0;
        this.idSIM = 0;
        this.ave = 0;
        this.duademo = 0;
        this.yeyuno = 0;
        this.lleon = 0;
        this.ciegos = 0;
        this.tonsillas = 0;
        this.higados = 0;
        this.molleja = 0;
        this.proventriculo = 0;
        this.flgGradoLesion = false;
    }
}